processing/requirements.txt
connexion[flask]
connexion[uvicorn]
swagger-ui-bundle
pyyaml
APScheduler
requests
starlette